requirements.txt
# 板块次日涨跌预测系统 - 依赖库

# Web框架
flask>=2.3.0

# 数据处理
pandas>=1.5.0
numpy>=1.23.0

# 数据获取 (如需真实数据请安装)
# akshare>=1.12.0

# 机器学习
scikit-learn>=1.2.0

# 梯度提升模型 (至少安装一个)
lightgbm>=3.3.0
# xgboost>=1.7.0

# 可视化
matplotlib>=3.6.0
seaborn>=0.12.0

# 配置文件
pyyaml>=6.0

# 工具
tqdm>=4.65.0
